Move initialization of vendor keys, app bins, ... to commands
Instead of initializing all of:

- vendor keys
- app binaries
- known firmwares

at start regardless of what command is given on the command line, we
do only the initialization we need for the specific command given.

This also means we can't report what exactly everything is built with
without initializing so we introduce a new function
util.go:builtWith() which does this for us.

We also use the correct digest of the no-touch signer.
